Configure Amplify for React Native
